# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for Auto Anime Downloader daemon",
        "title": "Auto Anime Downloader API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8091",
    "basePath": "/api/v1",
    "paths": {
        "/anilist/search": {
            "get": {
                "description": "Searches AniList by term and flags which results can be added as standalone animes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "standalone"
                ],
                "summary": "Search animes on AniList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term (fewer than 3 characters returns an empty list)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Include animes that have not aired yet (NOT_YET_RELEASED). Default false",
                        "name": "include_unreleased",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.AniListSearchResult"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    }
                }
            }
        },
...
        "/animes/{id}/episodes": {
            "get": {
                "description": "Returns anime info from AniList combined with downloaded episodes data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animes"
                ],
                "summary": "Get detail and episodes for a specific anime",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Anime ID (AniList media ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.AnimeDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AniListSearchResult": {
            "type": "object",
            "properties": {
                "block_reason": {
                    "description": "BlockReason e \"\" quando o anime pode ser adicionado como avulso, senao um dos motivos de\nstandalone_guard.go. E UM CAMPO, nao quatro booleanos: os motivos sao mutuamente\nexclusivos por precedencia e o card precisa de um rotulo so. Vem do mesmo blockReason que\no POST usa para decidir o 409 — e o que faz front e back concordarem por construcao.",
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "episodes": {
                    "description": "Episodes e 0 quando a AniList nao sabe o total (o comum em anime ainda no ar).",
                    "type": "integer",
                    "example": 24
                },
                "format": {
                    "type": "string",
                    "example": "TV"
                },
                "id": {
                    "type": "integer",
                    "example": 21
                },
                "status": {
                    "type": "string",
                    "example": "RELEASING"
                },
                "title": {
                    "type": "string",
                    "example": "One Piece"
                },
                "year": {
                    "type": "integer",
                    "example": 1999
                }
            }
        },
        "api.AnimeDetailResponse": {
            "type": "object",
            "properties": {
                "anime_id": {
                    "description": "AnimeID e o id da MIDIA na AniList — a mesma chave usada em episodes.json e nas rotas\n/animes/{id}/*, e tambem o id do link anilist.co/anime/{id} (ver decisions.md #43).",
                    "type": "integer"
                },
                "cover_image": {
                    "type": "string"
                },
                "custom_search_query": {
                    "type": "string"
                },
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.AnimeEpisodeInfo"
                    }
                },
                "progress": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_episodes": {
                    "type": "integer"
                }
            }
        },
        "api.AnimeEpisodeInfo": {
            "type": "object",
            "properties": {
                "airing_at": {
                    "type": "integer"
                },
                "download_date": {
                    "type": "string"
                },
                "episode_hash": {
                    "description": "EpisodeHash is the saved episode's torrent info hash. It is what lets the anime detail\nscreen join in live torrent progress exactly: a batch torrent's saved episodes all carry\nthe same hash but episode_number is meaningless for matching a torrent to them (the\ntorrent itself has no single episode number), so joining by number would miss batches\nentirely. Joining by hash instead, single and batch episodes use the same path. Empty\n(and omitted via omitempty) when the episode has no saved record.",
                    "type": "string"
                },
                "episode_name": {
                    "type": "string"
                },
                "episode_number": {
                    "description": "EpisodeNumber e a identidade do episodio nas rotas /animes/{id}/episodes/{episodeNumber}/*:\no id de no da AniList saiu de cena porque nao existe para episodio fora da janela de agenda\nque ela guarda (ver decisions.md #52).",
                    "type": "integer"
                },
                "is_aired": {
                    "type": "boolean"
                },
                "is_blocked": {
                    "type": "boolean"
                },
                "is_downloaded": {
                    "type": "boolean"
                },
                "is_manually_managed": {
                    "type": "boolean"
                },
                "is_watched": {
                    "type": "boolean"
                },
                "time_until_airing": {
                    "type": "integer"
                }
            }
        },
...
        "api.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/api.ErrorInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}